Crate network_simple_onchain_discovery

Structs

ConfigurationChangeListener
Listener which converts published updates from the OnChainConfig to ConnectivityRequests for the ConnectivityManager.

Statics

EVENT_PROCESSING_LOOP_BUSY_DURATION_S
Histogram of busy time of spent in event processing loop

EVENT_PROCESSING_LOOP_IDLE_DURATION_S
Histogram of idle time of spent in event processing loop